pyproject.toml
[project]
name = "aiavatar4discord"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord>=2.3.2",
    "google-adk>=1.16.0",
    "pyyaml>=6.0.3",
]
